refactor: bot events
