--- src/events/mod.rs
@@ -1,6 +1,0 @@
-pub mod anti_spam;
-pub mod daily_challenge;
-pub mod join;
-pub mod new_members_mention;
-pub mod send_message;
-pub mod read_github_links;
